+++ hack/pcr-reader/main_test.go
@@ -13,6 +13,7 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/edgelesssys/constellation/v2/internal/attestation/measurements"
 	"github.com/edgelesssys/constellation/v2/internal/attestation/vtpm"
 	"github.com/google/go-tpm-tools/proto/attest"
 	"github.com/google/go-tpm-tools/proto/tpm"
@@ -31,12 +32,12 @@ func TestMain(m *testing.M) {
 
 func TestExportToFile(t *testing.T) {
 	testCases := map[string]struct {
-		pcrs    map[uint32][]byte
+		pcrs    measurements.M
 		fs      *afero.Afero
 		wantErr bool
 	}{
 		"file not writeable": {
-			pcrs: map[uint32][]byte{
+			pcrs: measurements.M{
 				0: {0x1, 0x2, 0x3},
 				1: {0x1, 0x2, 0x3},
 				2: {0x1, 0x2, 0x3},
@@ -45,7 +46,7 @@ func TestExportToFile(t *testing.T) {
 			wantErr: true,
 		},
 		"file writeable": {
-			pcrs: map[uint32][]byte{
+			pcrs: measurements.M{
 				0: {0x1, 0x2, 0x3},
 				1: {0x1, 0x2, 0x3},
 				2: {0x1, 0x2, 0x3},
@@ -105,7 +106,7 @@ func TestValidatePCRAttDoc(t *testing.T) {
 						{
 							Pcrs: &tpm.PCRs{
 								Hash: tpm.HashAlgo_SHA256,
-								Pcrs: map[uint32][]byte{
+								Pcrs: measurements.M{
 									0: {0x1, 0x2, 0x3},
 								},
 							},
@@ -122,8 +123,8 @@ func TestValidatePCRAttDoc(t *testing.T) {
 						{
 							Pcrs: &tpm.PCRs{
 								Hash: tpm.HashAlgo_SHA256,
-								Pcrs: map[uint32][]byte{
-									0: []byte("AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"),
+								Pcrs: measurements.M{
+									0: measurements.PCRWithAllBytes(0xAA),
 								},
 							},
 						},
@@ -163,11 +164,11 @@ func mustMarshalAttDoc(t *testing.T, attDoc vtpm.AttestationDocument) []byte {
 
 func TestPrintPCRs(t *testing.T) {
 	testCases := map[string]struct {
-		pcrs   map[uint32][]byte
+		pcrs   measurements.M
 		format string
 	}{
 		"json": {
-			pcrs: map[uint32][]byte{
+			pcrs: measurements.M{
 				0: {0x1, 0x2, 0x3},
 				1: {0x1, 0x2, 0x3},
 				2: {0x1, 0x2, 0x3},
@@ -175,7 +176,7 @@ func TestPrintPCRs(t *testing.T) {
 			format: "json",
 		},
 		"empty format": {
-			pcrs: map[uint32][]byte{
+			pcrs: measurements.M{
 				0: {0x1, 0x2, 0x3},
 				1: {0x1, 0x2, 0x3},
 				2: {0x1, 0x2, 0x3},
@@ -183,7 +184,7 @@ func TestPrintPCRs(t *testing.T) {
 			format: "",
 		},
 		"yaml": {
-			pcrs: map[uint32][]byte{
+			pcrs: measurements.M{
 				0: {0x1, 0x2, 0x3},
 				1: {0x1, 0x2, 0x3},
 				2: {0x1, 0x2, 0x3},