+++ internal/attestation/measurements/measurements_test.go
@@ -4,7 +4,7 @@ Copyright (c) Edgeless Systems GmbH
 SPDX-License-Identifier: AGPL-3.0-only
 */
 
-package config
+package measurements
 
 import (
 	"context"
@@ -21,11 +21,11 @@ import (
 
 func TestMarshalYAML(t *testing.T) {
 	testCases := map[string]struct {
-		measurements  Measurements
+		measurements  M
 		wantBase64Map map[uint32]string
 	}{
 		"valid measurements": {
-			measurements: Measurements{
+			measurements: M{
 				2: []byte{253, 93, 233, 223, 53, 14, 59, 196, 65, 10, 192, 107, 191, 229, 204, 222, 185, 63, 83, 185, 239, 81, 35, 159, 117, 44, 230, 157, 188, 96, 15, 53},
 				3: []byte{213, 164, 73, 109, 33, 222, 201, 165, 37, 141, 219, 25, 198, 254, 181, 59, 180, 211, 192, 70, 63, 230, 7, 242, 72, 141, 223, 79, 16, 6, 239, 158},
 			},
@@ -35,7 +35,7 @@ func TestMarshalYAML(t *testing.T) {
 			},
 		},
 		"omit bytes": {
-			measurements: Measurements{
+			measurements: M{
 				2: []byte{},
 				3: []byte{1, 2, 3, 4},
 			},
@@ -63,7 +63,7 @@ func TestUnmarshalYAML(t *testing.T) {
 	testCases := map[string]struct {
 		inputBase64Map      map[uint32]string
 		forceUnmarshalError bool
-		wantMeasurements    Measurements
+		wantMeasurements    M
 		wantErr             bool
 	}{
 		"valid measurements": {
@@ -71,7 +71,7 @@ func TestUnmarshalYAML(t *testing.T) {
 				2: "/V3p3zUOO8RBCsBrv+XM3rk/U7nvUSOfdSzmnbxgDzU=",
 				3: "1aRJbSHeyaUljdsZxv61O7TTwEY/5gfySI3fTxAG754=",
 			},
-			wantMeasurements: Measurements{
+			wantMeasurements: M{
 				2: []byte{253, 93, 233, 223, 53, 14, 59, 196, 65, 10, 192, 107, 191, 229, 204, 222, 185, 63, 83, 185, 239, 81, 35, 159, 117, 44, 230, 157, 188, 96, 15, 53},
 				3: []byte{213, 164, 73, 109, 33, 222, 201, 165, 37, 141, 219, 25, 198, 254, 181, 59, 180, 211, 192, 70, 63, 230, 7, 242, 72, 141, 223, 79, 16, 6, 239, 158},
 			},
@@ -81,7 +81,7 @@ func TestUnmarshalYAML(t *testing.T) {
 				2: "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=",
 				3: "AQIDBAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=",
 			},
-			wantMeasurements: Measurements{
+			wantMeasurements: M{
 				2: []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
 				3: []byte{1, 2, 3, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
 			},
@@ -91,7 +91,7 @@ func TestUnmarshalYAML(t *testing.T) {
 				2: "This is not base64",
 				3: "AQIDBAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=",
 			},
-			wantMeasurements: Measurements{
+			wantMeasurements: M{
 				2: []byte{},
 				3: []byte{1, 2, 3, 4},
 			},
@@ -103,7 +103,7 @@ func TestUnmarshalYAML(t *testing.T) {
 				3: "AQIDBAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=",
 			},
 			forceUnmarshalError: true,
-			wantMeasurements: Measurements{
+			wantMeasurements: M{
 				2: []byte{},
 				3: []byte{1, 2, 3, 4},
 			},
@@ -116,7 +116,7 @@ func TestUnmarshalYAML(t *testing.T) {
 			assert := assert.New(t)
 			require := require.New(t)
 
-			var m Measurements
+			var m M
 			err := m.UnmarshalYAML(func(i any) error {
 				if base64Map, ok := i.(map[uint32]string); ok {
 					for key, value := range tc.inputBase64Map {
@@ -141,55 +141,55 @@ func TestUnmarshalYAML(t *testing.T) {
 
 func TestMeasurementsCopyFrom(t *testing.T) {
 	testCases := map[string]struct {
-		current          Measurements
-		newMeasurements  Measurements
-		wantMeasurements Measurements
+		current          M
+		newMeasurements  M
+		wantMeasurements M
 	}{
 		"add to empty": {
-			current: Measurements{},
-			newMeasurements: Measurements{
-				1: []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-				2: []byte{1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
-				3: []byte{2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2},
+			current: M{},
+			newMeasurements: M{
+				1: PCRWithAllBytes(0x00),
+				2: PCRWithAllBytes(0x01),
+				3: PCRWithAllBytes(0x02),
 			},
-			wantMeasurements: Measurements{
-				1: []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-				2: []byte{1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
-				3: []byte{2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2},
+			wantMeasurements: M{
+				1: PCRWithAllBytes(0x00),
+				2: PCRWithAllBytes(0x01),
+				3: PCRWithAllBytes(0x02),
 			},
 		},
 		"keep existing": {
-			current: Measurements{
-				4: []byte{1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
-				5: []byte{2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2},
+			current: M{
+				4: PCRWithAllBytes(0x01),
+				5: PCRWithAllBytes(0x02),
 			},
-			newMeasurements: Measurements{
-				1: []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-				2: []byte{1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
-				3: []byte{2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2},
+			newMeasurements: M{
+				1: PCRWithAllBytes(0x00),
+				2: PCRWithAllBytes(0x01),
+				3: PCRWithAllBytes(0x02),
 			},
-			wantMeasurements: Measurements{
-				1: []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-				2: []byte{1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
-				3: []byte{2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2},
-				4: []byte{1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
-				5: []byte{2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2},
+			wantMeasurements: M{
+				1: PCRWithAllBytes(0x00),
+				2: PCRWithAllBytes(0x01),
+				3: PCRWithAllBytes(0x02),
+				4: PCRWithAllBytes(0x01),
+				5: PCRWithAllBytes(0x02),
 			},
 		},
 		"overwrite existing": {
-			current: Measurements{
-				2: []byte{4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4},
-				3: []byte{5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5},
+			current: M{
+				2: PCRWithAllBytes(0x04),
+				3: PCRWithAllBytes(0x05),
 			},
-			newMeasurements: Measurements{
-				1: []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-				2: []byte{1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
-				3: []byte{2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2},
+			newMeasurements: M{
+				1: PCRWithAllBytes(0x00),
+				2: PCRWithAllBytes(0x01),
+				3: PCRWithAllBytes(0x02),
 			},
-			wantMeasurements: Measurements{
-				1: []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-				2: []byte{1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
-				3: []byte{2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2},
+			wantMeasurements: M{
+				1: PCRWithAllBytes(0x00),
+				2: PCRWithAllBytes(0x01),
+				3: PCRWithAllBytes(0x02),
 			},
 		},
 	}
@@ -230,7 +230,7 @@ func TestMeasurementsFetchAndVerify(t *testing.T) {
 		signature          string
 		signatureStatus    int
 		publicKey          []byte
-		wantMeasurements   Measurements
+		wantMeasurements   M
 		wantSHA            string
 		wantError          bool
 	}{
@@ -240,8 +240,8 @@ func TestMeasurementsFetchAndVerify(t *testing.T) {
 			signature:          "MEUCIBs1g2/n0FsgPfJ+0uLD5TaunGhxwDcQcUGBroejKvg3AiEAzZtcLU9O6IiVhxB8tBS+ty6MXoPNwL8WRWMzyr35eKI=",
 			signatureStatus:    http.StatusOK,
 			publicKey:          []byte("-----BEGIN PUBLIC KEY-----\nMFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEUs5fDUIz9aiwrfr8BK4VjN7jE6sl\ngz7UuXsOin8+dB0SGrbNHy7TJToa2fAiIKPVLTOfvY75DqRAtffhO1fpBA==\n-----END PUBLIC KEY-----"),
-			wantMeasurements: Measurements{
-				0: []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
+			wantMeasurements: M{
+				0: PCRWithAllBytes(0x00),
 			},
 			wantSHA: "4cd9d6ed8d9322150dff7738994c5e2fabff35f3bae6f5c993412d13249a5e87",
 		},
@@ -308,7 +308,7 @@ func TestMeasurementsFetchAndVerify(t *testing.T) {
 				}
 			})
 
-			m := Measurements{}
+			m := M{}
 			hash, err := m.FetchAndVerify(context.Background(), client, measurementsURL, signatureURL, tc.publicKey)
 
 			if tc.wantError {
@@ -318,6 +318,86 @@ func TestMeasurementsFetchAndVerify(t *testing.T) {
 			assert.Equal(tc.wantSHA, hash)
 			assert.NoError(err)
 			assert.EqualValues(tc.wantMeasurements, m)
+		})
+	}
+}
+
+func TestPCRWithAllBytes(t *testing.T) {
+	testCases := map[string]struct {
+		b       byte
+		wantPCR []byte
+	}{
+		"0x00": {
+			b:       0x00,
+			wantPCR: []byte{0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00},
+		},
+		"0x01": {
+			b:       0x01,
+			wantPCR: []byte{0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01},
+		},
+		"0xFF": {
+			b:       0xFF,
+			wantPCR: []byte{0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF},
+		},
+	}
+
+	for name, tc := range testCases {
+		t.Run(name, func(t *testing.T) {
+			assert := assert.New(t)
+			pcr := PCRWithAllBytes(tc.b)
+			assert.Equal(tc.wantPCR, pcr)
+		})
+	}
+}
+
+func TestEqualTo(t *testing.T) {
+	testCases := map[string]struct {
+		given     M
+		other     M
+		wantEqual bool
+	}{
+		"same values": {
+			given: M{
+				0: PCRWithAllBytes(0x00),
+				1: PCRWithAllBytes(0xFF),
+			},
+			other: M{
+				0: PCRWithAllBytes(0x00),
+				1: PCRWithAllBytes(0xFF),
+			},
+			wantEqual: true,
+		},
+		"different number of elements": {
+			given: M{
+				0: PCRWithAllBytes(0x00),
+				1: PCRWithAllBytes(0xFF),
+			},
+			other: M{
+				0: PCRWithAllBytes(0x00),
+			},
+			wantEqual: false,
+		},
+		"different values": {
+			given: M{
+				0: PCRWithAllBytes(0x00),
+				1: PCRWithAllBytes(0xFF),
+			},
+			other: M{
+				0: PCRWithAllBytes(0xFF),
+				1: PCRWithAllBytes(0x00),
+			},
+			wantEqual: false,
+		},
+	}
+
+	for name, tc := range testCases {
+		t.Run(name, func(t *testing.T) {
+			assert := assert.New(t)
+			if tc.wantEqual {
+				assert.True(tc.given.EqualTo(tc.other))
+			} else {
+				assert.False(tc.given.EqualTo(tc.other))
+			}
 		})
 	}
 }